Removed duplicate analytics Moved cookie control and analytics into _app.js
